[AIFlow][Refactor] Move server_runner and server_config to endpoint/server
