+++ flink-ai-flow/ai_flow/endpoint/server/aiflow-conf.yaml
@@ -16,6 +16,7 @@
 ## specific language governing permissions and limitations
 ## under the License.
 ##
+
 master_ip: localhost
 master_port: 50052
 db_uri: sqlite:///aiflow.db
