+++ flink-ai-flow/ai_flow/test/project/test_project.py
@@ -18,7 +18,7 @@
 #
 import unittest
 
-from ai_flow.application_master.server_runner import AIFlowServerRunner
+from ai_flow.endpoint.server.server_runner import AIFlowServerRunner
 
 from ai_flow.util.path_util import get_file_dir
 from ai_flow.project.project_description import get_project_description_from
