+++ flink-ai-flow/python_ai_flow/test/python_codes/evaluate_component/test_python_evaluate_component.py
@@ -27,7 +27,7 @@ from streamz import Stream
 from ai_flow.util.path_util import get_file_dir
 from ai_flow.executor.executor import PythonObjectExecutor
 from ai_flow.meta.model_meta import ModelType, ModelMeta
-from ai_flow.application_master.server_runner import AIFlowServerRunner
+from ai_flow.endpoint.server.server_runner import AIFlowServerRunner
 import ai_flow as af
 from ai_flow.udf.function_context import FunctionContext
 from ai_flow.meta.example_meta import ExampleSupportType, ExampleMeta
