+++ flink-ai-flow/python_ai_flow/test/python_codes/predict_component/test_python_predict_component.py
@@ -28,7 +28,7 @@ from streamz import Stream
 from ai_flow.meta.model_meta import ModelType, ModelMeta
 from ai_flow.executor.executor import PythonObjectExecutor
 from ai_flow.meta.example_meta import ExampleSupportType, ExampleMeta
-from ai_flow.application_master.server_runner import AIFlowServerRunner
+from ai_flow.endpoint.server.server_runner import AIFlowServerRunner
 from ai_flow.udf.function_context import FunctionContext
 from python_ai_flow import ExampleExecutor, Executor
 from python_ai_flow.test import test_util
