+++ flink-ai-flow/python_ai_flow/test/python_codes/push_component/test_python_push_component.py
@@ -22,12 +22,10 @@ import time
 import unittest
 from distutils.dir_util import copy_tree
 from typing import List
-from ai_flow import ExecutionMode
 from ai_flow.meta.example_meta import ExampleSupportType
 from ai_flow.executor.executor import PythonObjectExecutor
 from ai_flow.meta.model_meta import ModelMeta, ModelType
-from ai_flow.meta.artifact_meta import ArtifactMeta
-from ai_flow.application_master.server_runner import AIFlowServerRunner
+from ai_flow.endpoint.server.server_runner import AIFlowServerRunner
 from ai_flow.model_center.entity.model_version_stage import ModelVersionStage
 from ai_flow.udf.function_context import FunctionContext
 from python_ai_flow import Executor, ExampleExecutor
